@@ -1,16 +1,10 @@
 ﻿namespace Sim80C51.Processors.Attributes
 {
     [AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-    public class SFRAttribute : Attribute
+    public class SFRAttribute(ushort address, byte resetValue) : Attribute
     {
-        public ushort Address { get; }
+        public ushort Address => address;
 
-        public byte ResetValue { get; }
-
-        public SFRAttribute(ushort address, byte resetValue)
-        {
-            Address = address;
-            ResetValue = resetValue;
-        }
+        public byte ResetValue => resetValue;
     }
 }
