@@ -2,14 +2,10 @@
 
 namespace Sim80C51.Toolbox
 {
-    public sealed class ObservableKeyValuePair<TKey, TValue>: NotifyPropertyChanged
+    public sealed class ObservableKeyValuePair<TKey, TValue>(TKey key, TValue _value) : NotifyPropertyChanged
     {
-        public TKey Key { get { return _key; } set { _key = value; DoPropertyChanged(); } }
-        private TKey _key;
+        public TKey Key { get { return key; } set { key = value; DoPropertyChanged(); } }
 
         public TValue Value { get { return _value; } set { _value = value; DoPropertyChanged(); } }
-        private TValue _value;
-
-        public ObservableKeyValuePair(TKey key, TValue value) { _key = key; _value = value; }
     }
 }
