@@ -13,10 +13,10 @@ namespace Sim80C51.WSpace
 
         [YamlMember(ScalarStyle = ScalarStyle.Literal)]
         public string Listing { get; set; } = string.Empty;
-        public Dictionary<ushort, XMemConfig> XMem { get; set; } = new();
-        public List<ushort> Breakpoints { get; set; } = new();
-        public Dictionary<string, object> AdditionalSettings { get; set; } = new();
-        public Dictionary<ushort, string> MemoryWatches { get; set; } = new();
-        public List<ICallStackEntry> CallStack { get; set; } = new();
+        public Dictionary<ushort, XMemConfig> XMem { get; set; } = [];
+        public List<ushort> Breakpoints { get; set; } = [];
+        public Dictionary<string, object> AdditionalSettings { get; set; } = [];
+        public Dictionary<ushort, string> MemoryWatches { get; set; } = [];
+        public List<ICallStackEntry> CallStack { get; set; } = [];
     }
 }
